Add new dependencies to package-lock.json
This commit adds two new dependencies, "@actions/exec" and "@actions/glob", to the package-lock.json file. The version of "@actions/exec" is updated to 1.1.1 and it has a dependency on "@actions/io". The version of "@actions/glob" is set to 0.4.0 and it has dependencies on "@actions/core" (updated to 1.9.1) and "minimatch".